% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterhist.R
\name{scatterhist}
\alias{scatterhist}
\title{Title}
\usage{
scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{variable as vector}

\item{y}{variable as vector}

\item{xlab}{a label}

\item{ylab}{y label}
}
\value{
A sophesticated plot
}
\description{
Title
}
\examples{
\dontrun{with(ddt, scatterhist(LENGTH,WEIGHT, xlab="LENGTH"))}

}
